requirements.txt
gspread==6.0.0
google-auth==2.34.0
python-dotenv==1.0.1
requests
beautifulsoup4
google-generativeai>=0.7.2
feedparser>=6.0.11
python-dateutil>=2.9.0